+++ CHANGELOG.md
@@ -1,13 +1,20 @@
 # Change Log
 
+## [0.1.8] - 2017-02-27
+### Added
+- Added payment_method and payment_method_type in order class.
+- Added implementation for Wallet create, create_and_authenticate, authenticate, refresh_by_wallet_id, link & delink.
+- Added implementation for get_payment_methods in Payments class.
+- Added request parameters in error response.
+- Added api version as 2016-07-19.
 
-## [1.0.7] - 2016-11-10
+## [0.1.7] - 2016-11-10
 ### Added
 - PaymentLinks.java class, which contains the payment links for an order.
 - Order uuid in order creation response with 'id' as field.
 - Runtime dependency for unirest.
 
-## [1.0.6] - 2016-08-24
+## [0.1.6] - 2016-08-24
 ### Added
 - API implementation for Customer create, update, and get.
 - API implementation for Wallet list and refresh.